now the software supports the autonomous driving with force feedback (it needs a variable to change into this state in drive control)

+++ Inc/rotation_manager.h
@@ -25,7 +25,7 @@ Rotation_Manager_StatusTypeDef rotation_manager_init(
 		rotation_manager_t *rotation_manager, pid_t *pid,urb_sender_t *urb_sender);
 
 Rotation_Manager_StatusTypeDef rotation_manager_update(
-		rotation_manager_t *rotation_manager, int16_t auto_steer_feedback,
-		int16_t auto_control_steer);
+		rotation_manager_t *rotation_manager, double auto_steer_feedback,
+		double auto_control_steer);
 
 #endif /* INC_ROTATION_MANAGER_H_ */
